impriving flickering

+++ src/pseudo_rand.rs
@@ -1,5 +1,3 @@
-#![no_std]
-
 pub struct XorShift8 {
     state: i8,
 }
@@ -17,6 +15,6 @@ impl XorShift8 {
     pub fn random_between(&mut self, min: i8, max: i8) -> i8 {
         let r = self.rand8();
         let range = max - min + 1;
-        min + ((r.abs() as i8) % range)
+        min + (r.abs() % range)
     }
 }
